requirements.txt
streamlit==1.24.0
torch>=1.13.1
ultralytics==8.0.207
opencv-python==4.6.0.66
numpy==1.23.5
Pillow==9.5.0